% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transmutefalmouth.r
\name{transmutefalmouth}
\alias{transmutefalmouth}
\title{Transmute Falmouth data from a long-format vector of factors to a wide-format data frame of vectors}
\usage{
transmutefalmouth(falmouth_dataframe, pivot_factor)
}
\arguments{
\item{falmouth_dataframe}{(data frame) A dataframe of the type supplied by the getfalmouth() function}

\item{pivot_factor}{(str) the factor to transmute. Must be one of the following: 'aspd', 'avdir', 'cond', 'temp', 'pres'}
}
\value{
A matrix-like data frame
}
\description{
\code{transmutefalmouth} Transmute Falmouth from a vector of factors to a data frame of vectors
}
\details{
This function transmutes Falmouth data from the long format supplied to by getfalmouth() to a wide format, matrix-like dataframe
The resulting wide format dataframe is useful with functions that require a matrix or matrix-like input. (Correlation matrices, etc.)
}
\examples{
require(RPostgreSQL)
<<<<<<< HEAD
#replace the below line with your connection information
#db <- dbConnect(PostgreSQL(), host="localhost", user="wkp_user", password="wkp_user", dbname="wkp_hrdb")
fal.hourly <- getfalmouth(db, '2012-06-01 00:00:00', '2012-12-01 00:00:00', dataset="rfalmouth_hourly")
head(fal.hourly)
str(fal.hourly)
hourly.aspd <- transmutefalmouth(fal.hourly, "aspd")
head(hourly.aspd)
str(hourly.aspd)
plot(hourly.aspd[,seq(2, length(hourly.aspd))], pch=20)

#correlation matrix
cor(hourly.aspd[,seq(2, length(hourly.aspd))], use="pairwise.complete.obs")
heatmap(cor(hourly.aspd[,seq(2, length(hourly.aspd))], use="pairwise.complete.obs"),  col = cm.colors(256), symm=T, margins=c(12,12))
=======
con <- dbConnect(PostgreSQL(), host="localhost", user="wkp_user", password="wkp_user", dbname="wkp_hrdb")
fal <- getfalmouth(con, "2012-06-01 00:00:00", "2012-12-01 00:00:00")
head(fal)
str(fal)
fal2 <- transmutefalmouth(f, "aspd")
head(fal2)
str(fal2)
plot(dfal[,seq(2, length(dfal))], pch=20)
>>>>>>> parent of 27c03ab... Modify rkori
}
\seealso{
Other res: \code{\link{getfalmouth}};
  \code{\link{matplotfalmouth}}
}

